fix listGroups calls on the production portal
Please, read this carefully.

We must call the production portal on its .gov.br address because if we call
it over on .leg.br, it'll always redirect us to .gov.br, but not only that.
It'll redirect us to an "escaped" version of the original URL. So if we call a
method with the raw value '%', escaped correctly as '%25', it'll return a
redirect with it escaped as '%2525'. Like this:

    GET http://edemocracia.leg.br/?a=%25
    returns 301 Location http://edemocracia.gov.br/?a=%2525

So yeah, it pretty much breaks our client. So we from now on will ever call it
on the right URL. That's it. We should even add some checks in our code.