proper implementation of GroupListActivity
The Activity itself, in the future, will probably not be bound to content
type. It will use different ListFragments to show different kinds of content.
This is the first step in that direction. Now we have a simple Activity which
is only responsible for adding fragments and responding to some events. The
GroupListFragment is the only fragment available to that Activity for now.
It implements a ListFragment that shows a list of Groups.

In this right revision, tho, GroupListFragment's underlying ListAdapter is not
finished yet, so the representation of each item is not OK yet. That will come
in the next episodes.