working implementation of ThreadListFragment
Now the GroupListActivity will replace its initial GroupListFragment with a
ThreadListFragment when a group is selected. We can implement the other
screens as Fragments which will all be placed inside the GroupListActivity.

We should even rename that activity to something more generic, since it's now
the only interactive Activity in the application.