quick and dirty fix for the credential lost bug
Just reimplemented CookieCredentials through an internal CookieManager.

The *authenticate* method now puts the cookies in the header and then extracts
the cookies from the response. This way, the *authenticate* method is *firing*
the request, which is not desired. Ideally the *processResponse* action should
be performed by other method. I think there's a TODO for that in the code.