another big rework of the client code
We now use a Client, Session, Service model. The Service implements method
construction, signatures and so on. It uses the Session to communicate with
the remote web service. The Session keeps all the state required to properly
communicate with the remote service (user credentials, tokens, etc). It uses
the Client to actually do the plumbing work.

There are still some stuff to be figured out (like how to move the AuthToken
code out of the Client), and names still suck (ApplicationSession, wtf). But I
think its better than it used to be. Also, now we can *generate* our Service
layer.