simplify fragment instantiation
I'm tired of passing ContentUris around. Since I don't know how to actually do
it, its hard, error prone and verbose.

Fuck Uris. From now on we're gonna put Parcelable on the bundles. I made all
the Content subclasses *automagically* parcelable through Gson. It may not be
the best solution, but it works and it's easy.

So everything's simpler and everyone's happier from now on.

If we ever implement ContentProviders in the future we can review this.