Fix thread list always showing the same thread.
I really don't know if Pairs compare to each other based on the values
they hold, like I want them to do in this code, but, you know, I really
don't care. This whole thing stinks from the bottom up, so whatever.
It's gonna be this way until we are able to properly architect the whole
thing.