Implement background tasks as a Service.
A Service that can run in the background, I must add.

But there's more:

Replace piority-jobqueue with Tape. With this we lost connectivity
checks and retry logic, but that seems easy to recreate. Tape may seem a
lot more verbose at first, but that's just because it's lightweight.
With some clever refactoring we should be able to have parallel task
execution and simpler task classes. Anyway, we don't need that right now.

Remove the EventBus completely. Seems like we can replace every single
EventBus usage with a local broadcasts or RxJava awesomeness.

The communication between the tasks and the UI was removed with the
EventBus, but we shold reimplement that soon.