Improve google services authorization process.
Now the auth process is triggered once the Youtube account is changed.

If the process is not completed before a video is submitted, the service will create a notification asking the user to authorize the application.

The UI needs MAJOR polishing, but that'll have to wait.

Also, Tape library was removed. We are already storing everything we need in the SQLite database anyway.