Remove SplashScreenActivity.
Currently, SplashScreenActivity is only loading the *User* object so it's
available to anyone who needs it. This sucks at a lot of levels, since we
can't guarantee that everyone who needs the User will have it, nor the User
will be fresh, etc. If the user opens the application through a notification
and is presented with the MessageListFragment, for example, how can we
guarantee that the User object is available?

So I decided that everyone who needs the User object will get it from the
repository.

In order to not make one request each time we need a User, we can use the
current caching APIs. Since we only have a memory cache right now, I think we
could implement a new Cache subclass that stores its data in the current
account's *user data.*

For that to work, we have to ensure that an Account exists before doing
anything. So I decided to move the *check if an account exists* from the
Operator to a plain Observable. This way we can do stuff like this:

    ```
    AccountUtils.getOrRetrieveAccount(this)
                    .flatMap(account -> repository.getUser()
                            // Cache the User in the current account!
                            .compose(UserCache.with(this, account).cache()))
                    .flatMap(user -> repository.getGroups(user.getCompanyId()))
                    [...]
    ```

This will guarantee the integrity of the data across the whole application, no
matter where the user starts from.