Refactor video upload code to broadcast progress.
The main reason being our need to notify upload progress both via Android notifications and in the message list UI component.

Now we have a class that "holds" in progress uploads and can act like a bridge between the background task uploading the file and a UI component that wants to display that file's upload progress in some way. This class is the VideoAttachmentUploader, which is a singleton that will exist throught the entire application lifetime.